Revert "패키지 구조 변경"
This reverts commit 0f3629ab4e43f1bbecc8f20e9fa637ade5cb1d27.
