--- src/main/java/uniqram/c1one/search/application/dto/SearchHistoryDto.java
@@ -1,9 +1,0 @@
-package uniqram.c1one.search.application.dto;
-
-import lombok.Getter;
-
-@Getter
-public class SearchHistoryDto {
-
-	List<>
-}
