--- src/main/java/uniqram/c1one/search/domain/entity/SearchType.java
@@ -1,7 +1,0 @@
-package uniqram.c1one.search.domain.entity;
-
-public enum SearchType {
-	
-	HashTagSearch, UserSearch
-	
-}
