+++ src/main/java/uniqram/c1one/search/dto/HashtagSearchResultDto.java
@@ -1,0 +1,9 @@
+package uniqram.c1one.search.dto;
+
+import lombok.Getter;
+
+
+@Getter
+public class HashtagSearchResultDto {
+
+}
