+++ src/main/java/uniqram/c1one/search/entity/SearchHistory.java
@@ -1,9 +1,7 @@
-package uniqram.c1one.search.domain.entity;
+package uniqram.c1one.search.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
@@ -25,7 +23,5 @@ public class SearchHistory extends BaseEntity {
 	private String result;
 	
 	// Enum 만들기
-	// Mysql 에 varchar 로 저장
-	@Enumerated(EnumType.STRING)
-	private SearchType searchType;
+	
 }
