--- src/main/java/uniqram/c1one/search/infrastructure/dao/UserSearchDao.java
@@ -1,5 +1,0 @@
-package uniqram.c1one.search.infrastructure.dao;
-
-public class UserSearchDao {
-
-}
