+++ src/main/java/uniqram/c1one/search/service/FindSomething.java
@@ -1,8 +1,8 @@
-package uniqram.c1one.search.domain.service;
+package uniqram.c1one.search.service;
 
 
 import java.util.Optional;
-import uniqram.c1one.search.application.dto.UserSearchResultDto;
+import uniqram.c1one.search.dto.UserSearchResultDto;
 
 public interface FindSomething {
