+++ src/main/java/uniqram/c1one/search/service/FindSomethingImpl.java
@@ -1,12 +1,11 @@
-package uniqram.c1one.search.application.service;
+package uniqram.c1one.search.service;
 
 import java.util.Optional;
 import org.springframework.stereotype.Service;
-import uniqram.c1one.search.application.dto.UserSearchResultDto;
-import uniqram.c1one.search.domain.service.FindSomething;
+import uniqram.c1one.search.dto.UserSearchResultDto;
 
 @Service
-public class FindSomethingImpl implements FindSomething {
+public class FindSomethingImpl implements FindSomething{
 	
 	@Override
 	public Optional<UserSearchResultDto> findUser(String name) {
